nbinspect refactor into subcommands
