+++ appyter/render/flask_app/production.py
@@ -32,7 +32,7 @@ def serve(app_path, **kwargs):
       fw.write(render_form_from_nbtemplate(env, nbtemplate))
     with tmp_fs.open('index.json', 'w') as fw:
       import json
-      from appyter.render.nbinspect import render_nbtemplate_json_from_nbtemplate
+      from appyter.render.nbinspect.nbtemplate_json import render_nbtemplate_json_from_nbtemplate
       json.dump(render_nbtemplate_json_from_nbtemplate(env, nbtemplate), fw)
     with tmp_fs.open('landing.html', 'w') as fw:
       env.get_template('landing.j2').stream(
