+++ appyter/render/flask_app/static.py
@@ -10,7 +10,7 @@ from appyter.render.flask_app.util import route_join_with_or_without_slash
 from appyter.context import get_appyter_directory
 from appyter.parse.nb import nb_from_ipynb_io
 from appyter.render.form import render_form_from_nbtemplate
-from appyter.render.nbinspect import render_nbtemplate_json_from_nbtemplate
+from appyter.render.nbinspect.nbtemplate_json import render_nbtemplate_json_from_nbtemplate
 from appyter.context import get_jinja2_env
